feat(dom): add toHaveDescription assertion and implement getAccessibleDescription helper
